OpsPrice ppm u64

+++ src/constructive/core_types/ops_price/ext/codec/ape/encode/error/encode_error.rs
@@ -3,7 +3,7 @@
 pub enum OpsPriceAPEEncodeError {
     /// `ops_price_ppm` is strictly less than `base_ops_price`.
     OpsPriceBelowBaseOpsPrice {
-        ops_price_ppm: u32,
-        base_ops_price: u32,
+        ops_price_ppm: u64,
+        base_ops_price: u64,
     },
 }
